Solaris compat.

+++ hcore/compat.cxx
@@ -76,7 +76,7 @@ char* strcasestr( char const* haystack, char const* needle )
 #if ! defined( HAVE_GNU_GETHOSTBYNAME_R )
 int gethostbyname_r( char const* a0, struct hostent* a1, char* a2, size_t a3, struct hostent** a4, int* a5 )
 	{
-	hostent* h = ::gethostbyname_r( a0, a1, a2, a3, a5 );
+	hostent* h = ::gethostbyname_r( a0, a1, a2, static_cast<int>( a3 ), a5 );
 	if ( a4 )
 		*a4 = h;
 	return ( h ? 0 : errno );
@@ -88,7 +88,7 @@ int gethostbyname_r( char const* a0, struct hostent* a1, char* a2, size_t a3, st
 #if ! defined( HAVE_GNU_GETHOSTBYADDR_R )
 int gethostbyaddr_r( void const* a0, int a1, int a2, struct hostent* a3, char* a4, size_t a5, struct hostent** a6, int* a7 )
 	{
-	hostent* h = ::gethostbyaddr_r( static_cast<char const*>( a0 ), a1, a2, a3, a4, a5, a7 );
+	hostent* h = ::gethostbyaddr_r( static_cast<char const*>( a0 ), a1, a2, a3, a4, static_cast<int>( a5 ), a7 );
 	if ( a6 )
 		*a6 = h;
 	return ( h ? 0 : errno );
